requirements.txt
celery==5.3.1
fastapi
redis==4.6.0
uvicorn==0.22.0
flower==2.0.0
pydantic-settings